refactored requests
